+++ api/src/main/java/pl/budgee/adapter/web/WebModels.java
@@ -6,10 +6,15 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.Value;
 import pl.budgee.domain.model.Audit;
+import pl.budgee.domain.model.Budget.BudgetId;
+import pl.budgee.domain.model.Income;
+import pl.budgee.domain.model.IncomeType;
 import pl.budgee.domain.model.User;
+import pl.budgee.domain.usecase.CreateIncome.CreateIncomeRequest;
 import pl.budgee.domain.usecase.CreateUser.CreateUserRequest;
 import pl.budgee.domain.usecase.UpdateUser.UpdateUserRequest;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public class WebModels {
@@ -27,6 +32,18 @@ public class WebModels {
     }
   }
 
+  @Value
+  static class IncomeDto {
+    UUID id;
+    BigDecimal amount;
+    IncomeType type;
+    String description;
+
+    static IncomeDto of(Income income) {
+      return new IncomeDto(income.id().value(), income.amount(), income.type(), income.description());
+    }
+  }
+
   @Data
   @NoArgsConstructor
   static class CreateUserDto {
@@ -41,6 +58,18 @@ public class WebModels {
 
     CreateUserRequest toRequest() {
       return new CreateUserRequest(firstName, lastName, username, password);
+    }
+  }
+
+  @Data
+  @NoArgsConstructor
+  static class CreateIncomeDto {
+    BigDecimal amount;
+    IncomeType type;
+    String description;
+
+    CreateIncomeRequest toRequest(BudgetId id) {
+      return new CreateIncomeRequest(id, amount, type, description);
     }
   }