+++ api/src/main/java/pl/budgee/domain/model/BudgetNotFoundException.java
@@ -1,0 +1,10 @@
+package pl.budgee.domain.model;
+
+import pl.budgee.domain.model.Budget.BudgetId;
+
+public class BudgetNotFoundException extends RuntimeException{
+
+  public BudgetNotFoundException(BudgetId id) {
+    super("Budget of ID [" + id.value() + "] not found");
+  }
+}