+++ api/src/main/java/pl/budgee/domain/port/IncomeRepository.java
@@ -1,4 +1,9 @@
 package pl.budgee.domain.port;
 
+import pl.budgee.domain.model.Budget.BudgetId;
+import pl.budgee.domain.model.Income;
+
 public interface IncomeRepository {
+
+  Income save(Income income);
 }